fix: fix double amount binary inconcistency bug by switching to amount in cents via integers

+++ src/main/java/com/P6/P6/service/AccountService.java
@@ -14,7 +14,7 @@ public interface AccountService {
     double getBalance(UserEntity user);
 
     @Transactional
-    void transferMoney(Integer senderId, String receiver, double amount, String description);
+    void transferMoney(Integer senderId, String receiver, int amount, String description);
 
     @Transactional
     void deposit(UserEntity user, double amount);
